+++ RELEASE_NOTES.md
@@ -3,3 +3,5 @@
 ## Tournaments
 
 - Simplify the criteria to include them into the form (4.1.0)
+- New tie-breaks: STD, TPN and RTNG (4.1.0)
+- Update tie-breaks BH, FB and SB dummy scores according to FIDE Handbook C07.16 (4.1.0)